requirements.txt
langchain-community
langchain-google-genai
langchain_core
streamlit